refactor: extract cdp-key-mappings.ts from cdp-dispatch.ts (coherence split)
Separate keyboard layout data and char-to-key resolution from CDP protocol
dispatch logic. charToKeyInfo was the top bug-risk function (risk 0.68, 72 LOC)
in a mixed-role file.
